abstract pong from main into app/game
